adding utility class to move all common functions and static properties
